% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colpal.R
\name{displaypal}
\alias{displaypal}
\title{displays selected palette}
\usage{
displaypal(name, fontsize = 1, alpha = 1)
}
\arguments{
\item{name}{name of the palette. Use \code{palinfo()} for details}

\item{fontsize}{Default 1}

\item{alpha}{Default 1}
}
\description{
displays selected color palette in a graphics window
}
\examples{
displaypal(name = "german")
}
