% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_american}
\alias{scale_color_american}
\title{american color palette from faltuicolors}
\usage{
scale_color_american(palette = c("american"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
american color palette from faltuicolors
}
