% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_aussie}
\alias{scale_color_aussie}
\title{aussie color palette from faltuicolors}
\usage{
scale_color_aussie(palette = c("aussie"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
aussie color palette from faltuicolors
}
