% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_british}
\alias{scale_color_british}
\title{british color palette from faltuicolors}
\usage{
scale_color_british(palette = c("british"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
british color palette from faltuicolors
}
