% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_canadian}
\alias{scale_color_canadian}
\title{canadian color palette from faltuicolors}
\usage{
scale_color_canadian(palette = c("canadian"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
canadian color palette from faltuicolors
}
