% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_chinese}
\alias{scale_color_chinese}
\title{chinese color palette from faltuicolors}
\usage{
scale_color_chinese(palette = c("chinese"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
chinese color palette from faltuicolors
}
