% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_dutch}
\alias{scale_color_dutch}
\title{dutch color palette from faltuicolors}
\usage{
scale_color_dutch(palette = c("dutch"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
dutch color palette from faltuicolors
}
