% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_flatuiv1}
\alias{scale_color_flatuiv1}
\title{flatUIv1 color palette from faltuicolors}
\usage{
scale_color_flatuiv1(palette = c("flatUIv1"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
flatUIv1 color palette from faltuicolors
}
