% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_french}
\alias{scale_color_french}
\title{french color palette from faltuicolors}
\usage{
scale_color_french(palette = c("french"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
french color palette from faltuicolors
}
