% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_german}
\alias{scale_color_german}
\title{German color palette from faltuicolors}
\usage{
scale_color_german(palette = c("german"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
German color palette from faltuicolors
}
