% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_indian}
\alias{scale_color_indian}
\title{indian color palette from faltuicolors}
\usage{
scale_color_indian(palette = c("indian"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
indian color palette from faltuicolors
}
