% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_materialui}
\alias{scale_color_materialui}
\title{materialui color palette from faltuicolors}
\usage{
scale_color_materialui(
  palette = c("materialui"),
  alpha = 1,
  random = FALSE,
  ...
)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
materialui color palette from faltuicolors
}
