% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_spanish}
\alias{scale_color_spanish}
\title{spanish color palette from faltuicolors}
\usage{
scale_color_spanish(palette = c("spanish"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
spanish color palette from faltuicolors
}
