% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_swedish}
\alias{scale_color_swedish}
\title{swedish color palette from faltuicolors}
\usage{
scale_color_swedish(palette = c("swedish"), alpha = 1, random = FALSE, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
swedish color palette from faltuicolors
}
