% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scales.R
\name{scale_color_turkish}
\alias{scale_color_turkish}
\title{turkish color palette from faltuicolors}
\usage{
scale_color_turkish(palette = c("turkish"), random = FALSE, alpha = 1, ...)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
turkish color palette from faltuicolors
}
